requirements.txt
annotated-types==0.7.0; python_version >= '3.8'
appnope==0.1.4; platform_system == 'Darwin'
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.5.0; python_version >= '3.7'
certifi==2024.7.4; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
comm==0.2.2; python_version >= '3.8'
dbt==1.0.0.38.11; python_version >= '3.8'
debugpy==1.8.5; python_version >= '3.8'
decorator==5.1.1; python_version >= '3.5'
dlt[sql_database]==0.5.3; python_version < '3.13' and python_full_version >= '3.8.1'
executing==2.0.1; python_version >= '3.5'
fsspec==2024.6.1; python_version >= '3.8'
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.43; python_version >= '3.7'
giturlparse==0.12.0; python_version >= '3.8'
google-api-core[grpc]==2.19.1; python_version >= '3.7'
google-auth==2.34.0; python_version >= '3.7'
google-cloud-secret-manager==2.20.2; python_version >= '3.7'
googleapis-common-protos[grpc]==1.63.2; python_version >= '3.7'
greenlet==3.0.3; python_version < '3.13' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
grpc-google-iam-v1==0.13.1; python_version >= '3.7'
grpcio==1.66.0
grpcio-status==1.66.0
hexbytes==1.2.1; python_version >= '3.8' and python_version < '4'
humanize==4.10.0; python_version >= '3.8'
idna==3.8; python_version >= '3.6'
ipykernel==6.29.5; python_version >= '3.8'
ipython==8.26.0; python_version >= '3.10'
jedi==0.19.1; python_version >= '3.6'
jsonpath-ng==1.6.1
jupyter-client==8.6.2; python_version >= '3.8'
jupyter-core==5.7.2; python_version >= '3.8'
makefun==1.15.4
matplotlib-inline==0.1.7; python_version >= '3.8'
nest-asyncio==1.6.0; python_version >= '3.5'
numpy==2.1.0; python_version >= '3.10'
orjson==3.10.7; platform_python_implementation != 'PyPy'
packaging==24.1; python_version >= '3.8'
parso==0.8.4; python_version >= '3.6'
pathvalidate==3.2.1; python_version >= '3.7'
pendulum==3.0.0; python_version >= '3.8'
pexpect==4.9.0; sys_platform != 'win32' and sys_platform != 'emscripten'
platformdirs==4.2.2; python_version >= '3.8'
ply==3.11
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
proto-plus==1.24.0; python_version >= '3.7'
protobuf==5.27.3; python_version >= '3.8'
psutil==6.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
psycopg2-binary==2.9.9; python_version >= '3.7'
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==17.0.0; python_version >= '3.8'
pyasn1==0.6.0; python_version >= '3.8'
pyasn1-modules==0.4.0; python_version >= '3.8'
pydantic==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.1
pyyaml==6.0.2; python_version >= '3.8'
pyzmq==26.2.0; python_version >= '3.7'
requests==2.32.3; python_version >= '3.8'
requirements-parser==0.11.0; python_version >= '3.8' and python_version < '4.0'
rsa==4.9; python_version >= '3.6' and python_version < '4'
semver==3.0.2; python_version >= '3.7'
setuptools==73.0.1; python_version >= '3.8'
simplejson==3.19.3; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1; python_version >= '3.7'
sqlalchemy==2.0.32; python_version >= '3.7'
stack-data==0.6.3
tenacity==9.0.0; python_version >= '3.8'
time-machine==2.15.0; implementation_name != 'pypy'
tomlkit==0.13.2; python_version >= '3.8'
tornado==6.4.1; python_version >= '3.8'
traitlets==5.14.3; python_version >= '3.8'
types-setuptools==73.0.0.20240822; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.1; python_version >= '2'
urllib3==2.2.2; python_version >= '3.8'
wcwidth==0.2.13
wheel==0.44.0; python_version >= '3.8'
